Conflicts:
	admin/inc/basic.php
	admin/theme-edit.php

+++ admin/inc/common.php
@@ -13,6 +13,8 @@
  * Bad stuff protection
  */
 define('IN_GS', TRUE);
+define('GSSTYLEWIDE','wide');
+
 include_once('security_functions.php');
 
 if (version_compare(PHP_VERSION, "5")  >= 0) {
@@ -25,7 +27,6 @@ if (version_compare(PHP_VERSION, "5")  >= 0) {
 include('basic.php');
 include('template_functions.php');
 include('logging.class.php');
-
 
 define('GSROOTPATH', get_root_path());
 
@@ -57,8 +58,6 @@ define('GSUSERSPATH', GSROOTPATH. 'data/users/');
 define('GSBACKUSERSPATH', GSROOTPATH. 'backups/users/');
 define('GSCACHEPATH', GSROOTPATH. 'data/cache/');
 define('GSAUTOSAVEPATH', GSROOTPATH. 'data/pages/autosave/');
-
-
 
 /**
  * Variable check to prevent debugging going off
@@ -189,21 +188,23 @@ if (get_filename_id() != 'install' && get_filename_id() != 'setup' && get_filena
 	if ($SITEURL == '')	{
 		redirect($fullpath . $GSADMIN.'/install.php');
 	} 
-	
-	# if you've made it this far, the site is already installed so remove the installation files
-	$filedeletionstatus=true;
-	if (file_exists(GSADMINPATH.'install.php'))	{
-		$filedeletionstatus = unlink(GSADMINPATH.'install.php');
-	}
-	if (file_exists(GSADMINPATH.'setup.php'))	{
-		$filedeletionstatus = unlink(GSADMINPATH.'setup.php');
-	}
-	if (file_exists(GSADMINPATH.'update.php'))	{
-		$filedeletionstatus = unlink(GSADMINPATH.'update.php');
-	}
-	if (!$filedeletionstatus) {
-		$error = sprintf(i18n_r('ERR_CANNOT_DELETE'), '<code>/'.$GSADMIN.'/install.php</code>, <code>/'.$GSADMIN.'/setup.php</code> or <code>/'.$GSADMIN.'/update.php</code>');
-	}
+
+	if(!getDef('GSDEBUGINSTALL',true)){	
+		# if you've made it this far, the site is already installed so remove the installation files
+		$filedeletionstatus=true;
+		if (file_exists(GSADMINPATH.'install.php'))	{
+			$filedeletionstatus = unlink(GSADMINPATH.'install.php');
+		}
+		if (file_exists(GSADMINPATH.'setup.php'))	{
+			$filedeletionstatus = unlink(GSADMINPATH.'setup.php');
+		}
+		if (file_exists(GSADMINPATH.'update.php'))	{
+			$filedeletionstatus = unlink(GSADMINPATH.'update.php');
+		}
+		if (!$filedeletionstatus) {
+			$error = sprintf(i18n_r('ERR_CANNOT_DELETE'), '<code>/'.$GSADMIN.'/install.php</code>, <code>/'.$GSADMIN.'/setup.php</code> or <code>/'.$GSADMIN.'/update.php</code>');
+		}
+	}	
 }
